fixed DummyFileHelper size type to long

+++ src/Downloader.DummyHttpServer/DummyFileHelper.cs
@@ -20,46 +20,46 @@ public static class DummyFileHelper
         HttpServer.Run(0); // with dynamic port
     }
 
-    public static string GetFileUrl(int size)
+    public static string GetFileUrl(long size)
     {
         return $"http://localhost:{Port}/dummyfile/file/size/{size}";
     }
 
-    public static string GetFileWithNameUrl(string filename, int size, byte? fillByte = null)
+    public static string GetFileWithNameUrl(string filename, long size, byte? fillByte = null)
     {
         return $"http://localhost:{Port}/dummyfile/file/{filename}?size={size}"
             + (fillByte == null ? "" : $"&fillByte={fillByte}");
     }
 
-    public static string GetFileWithNameOnRedirectUrl(string filename, int size)
+    public static string GetFileWithNameOnRedirectUrl(string filename, long size)
     {
         return $"http://localhost:{Port}/dummyfile/file/{filename}/redirect?size={size}";
     }
 
-    public static string GetFileWithoutHeaderUrl(string filename, int size, byte? fillByte = null)
+    public static string GetFileWithoutHeaderUrl(string filename, long size, byte? fillByte = null)
     {
         return $"http://localhost:{Port}/dummyfile/noheader/file/{filename}?size={size}"
             + (fillByte == null ? "" : $"&fillByte={fillByte}");
     }
 
-    public static string GetFileWithContentDispositionUrl(string filename, int size, byte? fillByte = null)
+    public static string GetFileWithContentDispositionUrl(string filename, long size, byte? fillByte = null)
     {
         return $"http://localhost:{Port}/dummyfile/file/{filename}/size/{size}"
             + (fillByte == null ? "" : $"?fillByte={fillByte}");
     }
 
-    public static string GetFileWithNoAcceptRangeUrl(string filename, int size, byte? fillByte = null)
+    public static string GetFileWithNoAcceptRangeUrl(string filename, long size, byte? fillByte = null)
     {
         return $"http://localhost:{Port}/dummyfile/file/{filename}/size/{size}/norange"
             + (fillByte == null ? "" : $"?fillByte={fillByte}");
     }
 
-    public static string GetFileWithFailureAfterOffset(int size, int failureOffset)
+    public static string GetFileWithFailureAfterOffset(long size, int failureOffset)
     {
         return $"http://localhost:{Port}/dummyfile/file/size/{size}/failure/{failureOffset}";
     }
 
-    public static string GetFileWithTimeoutAfterOffset(int size, int timeoutOffset)
+    public static string GetFileWithTimeoutAfterOffset(long size, int timeoutOffset)
     {
         return $"http://localhost:{Port}/dummyfile/file/size/{size}/timeout/{timeoutOffset}";
     }
